requirements.txt
Homie4
paho-mqtt~=1.6.1
APScheduler==3.10.0
service-bootstrap~=0.0.5
homie-helpers==0.0.7.dev1677103362
